+++ src/pylinkage/joints/fixed.py
@@ -2,17 +2,13 @@
 Fixed joint.
 """
 
-from __future__ import annotations
 
 import math
-from typing import TYPE_CHECKING
 
+from .._types import Coord
 from .. import exceptions as pl_exceptions
 from .. import geometry as pl_geom
 from . import joint as pl_joint
-
-if TYPE_CHECKING:
-    from .._types import Coord
 
 
 class Fixed(pl_joint.Joint):