+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Switches to `uv`.
 - Renames `HypostaticError` to `UnderconstrainedError` and `hyperstaticity()` to `indeterminacy()`.
   Old names kept as deprecated aliases.
+- Adds an internal solver with numba, almost 100x faster!
 
 ### Fixed